fix(app): 描画機会の番号を受ける境界を64ビットへ広げ、32ビットへの縮めを消す
描画機会の番号を64ビットへ広げたのに、番号を受ける3つの境界(画素内ずらしの列の添字・
スモークの計画・ストリーミングの固定経路と計測)が32ビットのままだったため、ストリーミングを
有効にした実行は毎描画で縮める算術を通り、約43億回目の描画で型付きエラーになって終了した。
64ビットへ広げた目的そのものを潰していたため、受け手をすべて64ビットで受ける形へ直す。

固定経路の往復の位相は距離で決まり、番号だけを整数の剰余で折り返せないため、番号を上位32ビットと
下位32ビットへ分けて実数へ写す(`u64`から`f64`への`From`は標準が備えない)。画素内ずらしの列の
添字は列の長さで折り返した余りを32ビットへ戻し、収まらない事態は到達したらバグとしてpanicにする。

縮めが消えたことで、`描画機会の番号`の縮める口と`起動エラー`の枝、視点の配線と自己操作の配線が
番号のために返していた`Result`も消える。型計測の台帳の`起動エラー`の枝数を実測の46へ戻す。

+++ crates/blitz_app/src/app/streaming/measurement.rs
@@ -50,8 +50,8 @@ impl ストリーミング計測 {
         }
     }
 
-    pub(super) fn 記録する(&mut self, フレーム番号: u32, 所要時間: Duration, 進行: &ストリーミング進行, 段選択: &[(blitz_engine::チャンク座標, blitz_render::地形詳細段)]) {
-        if フレーム番号 >= ウォームアップフレーム数 {
+    pub(super) fn 記録する(&mut self, フレーム番号: u64, 所要時間: Duration, 進行: &ストリーミング進行, 段選択: &[(blitz_engine::チャンク座標, blitz_render::地形詳細段)]) {
+        if フレーム番号 >= u64::from(ウォームアップフレーム数) {
             self.処理時間一覧ms.push(所要時間.as_secs_f64() * 1000.0);
         }
         self.観測フレーム数 = self.観測フレーム数.saturating_add(1);
